Added optional defines for URLs in uninstall info
Added the optional URL details for the uninstall info in the registry that would otherwise have to be manually coded into the installer script, which partially defeats the benefit of using the MULTIUSER_RegistryAddInstallInfo macro. The links will show up at the bottom of the Programs and Features window when selecting an application in the list to uninstall.