Workaround for Windows uninstaller elevation bug
When the uninstaller is started as administrator but not from the
installer, it restarts itself under the user of the running shell,
in order to overcome the Windows bugs that:
- Elevates the uninstallers of single-user installations when
called from 'Apps & features' of Windows 10 causing them to fail
when using a different account for elevation.
- Elevates the uninstallers of all-users installations when called
from 'Add/Remove Programs' of Control Panel, preventing them of
eleveting on their own and correctly recognize the user that started
the uninstaller. If a different account was used for elevation, all
user-context operations will be performed for the user of that
account. In this case, the fix causes the elevetion prompt to be
displayed twice (one from Control Panel and one from the uninstaller).
Closes #6